feat: import AsideNavbarData from data file
